Robust.Packaging updates
Port some stuff from python packaging.
