+++ Robust.Packaging/RobustClientPackaging.cs
@@ -4,9 +4,10 @@ namespace Robust.Packaging;
 
 public sealed class RobustClientPackaging
 {
-    public static IReadOnlySet<string> ClientIgnoresResources { get; } = new HashSet<string>
+    public static IReadOnlySet<string> ClientIgnoredResources { get; } = new HashSet<string>
     {
         "Maps",
+        "ConfigPresets",
         // Leaving this here for future archaeologists to ponder at.
         "emotes.xml",
         "Groups",
@@ -18,48 +19,8 @@ public sealed class RobustClientPackaging
         AssetPass pass,
         CancellationToken cancel = default)
     {
-        var ignoreSet = ClientIgnoresResources.Union(RobustSharedPackaging.SharedIgnoredResources).ToHashSet();
+        var ignoreSet = ClientIgnoredResources.Union(RobustSharedPackaging.SharedIgnoredResources).ToHashSet();
 
-        await RobustSharedPackaging.DoResourceCopy(Path.Combine(contentDir, "Resources"), pass, ignoreSet, cancel);
-    }
-
-    public static async Task WriteContentAssemblies(
-        AssetPass pass,
-        string contentDir,
-        string binDir,
-        IEnumerable<string> contentAssemblies,
-        CancellationToken cancel = default)
-    {
-        await WriteContentAssemblies("Assemblies", pass, contentDir, binDir, contentAssemblies, cancel);
-    }
-
-    public static Task WriteContentAssemblies(
-        string target,
-        AssetPass pass,
-        string contentDir,
-        string binDir,
-        IEnumerable<string> contentAssemblies,
-        CancellationToken cancel = default)
-    {
-        var files = new List<string>();
-
-        var sourceDir = Path.Combine(contentDir, "bin", binDir);
-
-        foreach (var asm in contentAssemblies)
-        {
-            files.Add($"{asm}.dll");
-
-            var pdbPath = $"{asm}.pdb";
-            if (File.Exists(Path.Combine(sourceDir, pdbPath)))
-                files.Add(pdbPath);
-        }
-
-        foreach (var f in files)
-        {
-            cancel.ThrowIfCancellationRequested();
-            pass.InjectFileFromDisk($"{target}/{f}", Path.Combine(sourceDir, f));
-        }
-
-        return Task.CompletedTask;
+        await RobustSharedPackaging.DoResourceCopy(Path.Combine(contentDir, "Resources"), pass, ignoreSet, cancel: cancel);
     }
 }
